# gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a chatting API",
        "title": "Jackey Chatting API",
        "contact": {
            "name": "API Support",
            "url": "http://www.example.com/support",
            "email": "support@example.com"
        },
        "version": "1"
    },
    "host": "localhost:8888",
    "basePath": "/api/",
    "paths": {
...
        "/user/profile": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a user's profile",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User/profile"
                ],
                "summary": "Get a user's profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handler._BaseResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.GetUserProfileResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/user/profile/avatar": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Download a user's avatar",
                "produces": [
                    "image/jpeg"
                ],
                "tags": [
                    "User/profile"
                ],
                "summary": "Download a user's avatar",
                "responses": {
                    "200": {
                        "description": "avatar",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/user/profile/avatar/upload": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload a user's avatar",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User/profile"
                ],
                "summary": "Upload a user's avatar",
                "parameters": [
                    {
                        "type": "file",
                        "description": "avatar",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler._BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
        "/user/profile/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update a user's profile",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User/profile"
                ],
                "summary": "Update a user's profile",
                "parameters": [
                    {
                        "description": "updateProfile request",
                        "name": "avatar",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UpdateUserProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler._BaseResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler._ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler._BaseResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "数据"
                },
                "status": {
                    "description": "状态",
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "handler._ResponseError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误信息",
                    "type": "string",
                    "example": "err msg"
                },
                "status": {
                    "description": "状态",
                    "type": "string",
                    "example": "failed"
                }
            }
        },
...
        "user.GetUserProfileResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
...
        "user.UpdateUserProfileRequest": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}